requirements.txt
flask-restful
Flask-OIDC
APISpec
marshmallow